move node state into its own dedicated module along with deposits
