+++ hydra-node/src/Hydra/HeadLogic/StateEvent.hs
@@ -5,8 +5,8 @@ module Hydra.HeadLogic.StateEvent where
 
 import Hydra.Chain.ChainState (IsChainState)
 import Hydra.Events (EventId, HasEventId (..))
-import Hydra.HeadLogic (NodeState)
 import Hydra.HeadLogic.Outcome (StateChanged (Checkpoint))
+import Hydra.Node.State (NodeState)
 import Hydra.Prelude
 import Hydra.Tx (ArbitraryIsTx)
